some restructure
